+++ src/components/molecules/ModalConfirmWithNamespaceSelect/ModalConfirmWithNamespaceSelect.tsx
@@ -9,10 +9,10 @@ import {currentClusterAccessSelector, kubeConfigPathSelector} from '@redux/selec
 
 import {useTargetClusterNamespaces} from '@hooks/useTargetClusterNamespaces';
 
-import {createKubeClient} from '@utils/kubeclient';
 import {getDefaultNamespaceForApply} from '@utils/resources';
 
 import {K8sResource} from '@shared/models/k8sResource';
+import {createKubeClient} from '@shared/utils/kubeclient';
 import {kubeConfigContextSelector} from '@shared/utils/selectors';
 
 import * as S from './ModalConfirmWithNamespaceSelect.styled';
