+++ src/constants/constants.ts
@@ -9,7 +9,6 @@ export const YAML_DOCUMENT_DELIMITER_NEW_LINE = '---\n';
 export const TOOLTIP_DELAY = 1.0;
 export const LONGER_TOOLTIP_DELAY = 2.0;
 export const TOOLTIP_K8S_SELECTION = 'Select which kubernetes schema version to use for validation';
-export const ERROR_MSG_FALLBACK = 'Looks like something unexpected went wrong. Please try again later.';
 export const REF_PATH_SEPARATOR = '#';
 export const KUSTOMIZATION_FILE_NAME = 'kustomization.yaml';
 export const KUSTOMIZATION_KIND = 'Kustomization';
