+++ src/redux/compare/utils.ts
@@ -1,6 +1,5 @@
-import {isDefined} from '@utils/filter';
-
 import {PartialResourceSet, ResourceSet} from '@shared/models/compare';
+import {isDefined} from '@shared/utils/filter';
 
 export function isCompleteResourceSet(options: PartialResourceSet | undefined): options is ResourceSet {
   switch (options?.type) {
