+++ src/redux/services/compare/fetchResources.ts
@@ -5,27 +5,15 @@ import path, {sep} from 'path';
 import invariant from 'tiny-invariant';
 import {v4 as uuid} from 'uuid';
 
-import {
-  CLUSTER_DIFF_PREFIX,
-  ERROR_MSG_FALLBACK,
-  PREVIEW_PREFIX,
-  YAML_DOCUMENT_DELIMITER_NEW_LINE,
-} from '@constants/constants';
+import {CLUSTER_DIFF_PREFIX, PREVIEW_PREFIX, YAML_DOCUMENT_DELIMITER_NEW_LINE} from '@constants/constants';
 
 import {currentConfigSelector, kubeConfigPathSelector} from '@redux/selectors';
 import {runKustomize} from '@redux/thunks/previewKustomization';
 
-import {
-  createHelmInstallCommand,
-  createHelmTemplateCommand,
-  hasCommandFailed,
-  runCommandInMainThread,
-} from '@utils/commands';
-import {isDefined} from '@utils/filter';
 import {buildHelmCommand} from '@utils/helm';
-import {createKubeClient} from '@utils/kubeclient';
 import {promiseFromIpcRenderer} from '@utils/promises';
 
+import {ERROR_MSG_FALLBACK} from '@shared/constants/constants';
 import {ROOT_FILE_ENTRY} from '@shared/constants/fileEntry';
 import {CommandOptions} from '@shared/models/commands';
 import {
@@ -40,6 +28,14 @@ import {
 } from '@shared/models/compare';
 import {K8sResource} from '@shared/models/k8sResource';
 import {RootState} from '@shared/models/rootState';
+import {
+  createHelmInstallCommand,
+  createHelmTemplateCommand,
+  hasCommandFailed,
+  runCommandInMainThread,
+} from '@shared/utils/commands';
+import {isDefined} from '@shared/utils/filter';
+import {createKubeClient} from '@shared/utils/kubeclient';
 
 import getClusterObjects from '../getClusterObjects';
 import {isKustomizationResource} from '../kustomize';
