+++ src/redux/services/compare/filterComparisons.ts
@@ -1,9 +1,8 @@
-import {isPassingKeyValueFilter} from '@utils/filter';
-
 import {getResourceKindHandler} from '@src/kindhandlers';
 
 import {CompareFilter, ResourceComparison} from '@shared/models/compare';
 import {K8sResource} from '@shared/models/k8sResource';
+import {isPassingKeyValueFilter} from '@shared/utils/filter';
 
 type FilterOptions = {
   search?: string;
