+++ src/redux/thunks/previewReferencedHelmChart.ts
@@ -7,11 +7,10 @@ import {extractObjectsFromYaml} from '@redux/services/manifest-utils';
 import {interpolateTemplate} from '@redux/services/templates';
 import {createUnsavedResource} from '@redux/services/unsavedResource';
 
-import {runCommandInMainThread} from '@utils/commands';
-
 import {AppDispatch} from '@shared/models/appDispatch';
 import {CommandOptions} from '@shared/models/commands';
 import {K8sResource} from '@shared/models/k8sResource';
+import {runCommandInMainThread} from '@shared/utils/commands';
 
 const fsWriteFilePromise = promisify(fs.writeFile);
 const fsReadFilePromise = promisify(fs.readFile);
