+++ src/shared/constants/constants.ts
@@ -1,0 +1,1 @@
+export const ERROR_MSG_FALLBACK = 'Looks like something unexpected went wrong. Please try again later.';
