+++ src/shared/utils/commands/kubectl.ts
