+++ src/utils/error.ts
@@ -1,4 +1,4 @@
-import {ERROR_MSG_FALLBACK} from '@constants/constants';
+import {ERROR_MSG_FALLBACK} from '@shared/constants/constants';
 
 export function errorMsg(err: unknown) {
   return err instanceof Error ? err.message : ERROR_MSG_FALLBACK;
